Update Cates_small_project_1.py
Now getting an "out of bounds" error on line 47 within the score "function". This occurred when I tested it with two sequences of different lengths. I think it has something to do with the ranges of the nested for loops  on lines 26 & 27.

The goal here is to compare all of sequence 2 (which is j) with a single base in sequence 1 (which is i) and move column-wise when populating the DP table.  So I'm thinking the ranges of my loops are somehow messed up....