Update Cates_small_project_1.py
Fixed the out of bounds error in lines 24-34 by changing the ranges of the outer and inner for loops in the "score" function. The outer loop now iterates through sequence 2 first (i, along the vertical of the table) and the inner loop iterates through sequence 1 (j, along the horizontal of the table). 
BUT, the alignment produced from the "traceback" function isn't correct, so I still need to figure that out first and then work on the "calculate_score" function once the alignment is correct.