Update Cates_small_project_1.py
Don't need to add up the score as you do the alignment. The max score in the table IS the final score since all the cells are cumulative. So, in the traceback function call in the score function, I added "final_max_val" as an argument just so I can call it in the final results printout